set permission in controller

+++ resources/views/backend/pages/attendance/index.blade.php
@@ -82,12 +82,14 @@
                                             </td>
                                             <td>
                                                 <a href="{{ route('attendances.edit', $attendance->id) }}" class="btn btn-primary">Edit</a>
-                                                <button type="button" class="btn btn-danger" onclick="deleteAttendances({{ $attendance->id }})">Delete</button>
+                                                @can('attendance delete')
+                                                    <button type="button" class="btn btn-danger" onclick="deleteAttendances({{ $attendance->id }})">Delete</button>
 
-                                                <form action="{{ route('attendances.destroy', $attendance->id) }}" id="delete-form-{{ $attendance->id }}" method="post" style="display: none">
-                                                    @csrf
-                                                    @method('DELETE')
-                                                </form>
+                                                    <form action="{{ route('attendances.destroy', $attendance->id) }}" id="delete-form-{{ $attendance->id }}" method="post" style="display: none">
+                                                        @csrf
+                                                        @method('DELETE')
+                                                    </form>
+                                                @endcan
 
                                             </td>
                                         </tr>
